Create README.md
PlayDots is an eLearning SaaS application based off of the game dots and boxes. There are two game modes (Standard and Create Your Own Trivia).

How to Play:
Two players will take turns joining two adjacent dots with a horizontal or vertical line. When a player completes the fourth side of a box they must answer a fun trivia question before the time on the clock runs out. When all the boxes have been completed the winner is the player with the most completed boxes.

Built With:
- HTML
- CSS
- JavaScript
- JQuery